+++ StripePaymentSheet/StripePaymentSheet/Source/PaymentSheet/PaymentSheet+API.swift
@@ -590,104 +590,6 @@ extension PaymentSheet {
         }
         return params
     }
-
-    private static func createPaymentMethodIfNeeded(apiClient: STPAPIClient,
-                                                    paymentMethod: STPPaymentMethod?,
-                                                    paymentMethodParams: STPPaymentMethodParams?) async throws -> STPPaymentMethod {
-        if let paymentMethod = paymentMethod {
-            return paymentMethod
-        }
-
-        guard let paymentMethodParams = paymentMethodParams else {
-            throw PaymentSheetError.unknown(debugDescription: "paymentMethodParams unexpectedly nil")
-        }
-
-        return try await apiClient.createPaymentMethod(with: paymentMethodParams)
-    }
-
-    private static func handleDeferredIntentConfirmation(deferredIntentContext: DeferredIntentContext,
-                                                         paymentMethod: STPPaymentMethod?,
-                                                         paymentMethodParams: STPPaymentMethodParams?,
-                                                         shouldSavePaymentMethod: Bool) {
-        Task {
-            do {
-                // Create PM if necessary
-                let _paymentMethod = try await createPaymentMethodIfNeeded(apiClient: deferredIntentContext.configuration.apiClient,
-                                                                           paymentMethod: paymentMethod,
-                                                                           paymentMethodParams: paymentMethodParams)
-                // Get Intent client secret from merchant
-                let clientSecret = try await fetchIntentClientSecretFromMerchant(intentConfig: deferredIntentContext.intentConfig,
-                                                                                 paymentMethodID: _paymentMethod.stripeId,
-                                                                                 shouldSavePaymentMethod: shouldSavePaymentMethod)
-                // Finish confirmation
-                if deferredIntentContext.isServerSideConfirmation {
-                    // Server-side confirmation
-                    func handleStatus(status: STPPaymentHandlerActionStatus, error: Error?) {
-                        switch status {
-                        case .succeeded:
-                            deferredIntentContext.completion(.completed)
-                        case .canceled:
-                            deferredIntentContext.completion(.canceled)
-                        case .failed:
-                            let error = error ?? PaymentSheetError.unknown(debugDescription: "Unknown error occured while handling intent next action")
-                            deferredIntentContext.completion(.failed(error: error))
-                        @unknown default:
-                            deferredIntentContext.completion(.failed(error: PaymentSheetError.unknown(debugDescription: "Unrecognized intent status")))
-                        }
-                    }
-
-                    switch deferredIntentContext.intentConfig.mode {
-                    case .payment:
-                        deferredIntentContext.paymentHandler.handleNextAction(forPayment: clientSecret,
-                                                                        with: deferredIntentContext.authenticationContext,
-                                                                        returnURL: deferredIntentContext.configuration.returnURL,
-                                                                        completion: { status, _, error in
-                            handleStatus(status: status, error: error)
-                        })
-                    case .setup:
-                        deferredIntentContext.paymentHandler.handleNextAction(forSetupIntent: clientSecret,
-                                                                        with: deferredIntentContext.authenticationContext,
-                                                                        returnURL: deferredIntentContext.configuration.returnURL,
-                                                                        completion: { status, _, error in
-                            handleStatus(status: status, error: error)
-                        })
-                    }
-                } else {
-                    // Client-side confirmation
-                    let intent = try await deferredIntentContext.configuration.apiClient.retrieveIntent(for: deferredIntentContext.intentConfig,
-                                                                                                        withClientSecret: clientSecret)
-                    confirm(configuration: deferredIntentContext.configuration,
-                            authenticationContext: deferredIntentContext.authenticationContext,
-                            intent: intent,
-                            paymentOption: deferredIntentContext.paymentOption,
-                            paymentHandler: deferredIntentContext.paymentHandler,
-                            completion: deferredIntentContext.completion)
-                }
-
-            } catch {
-                DispatchQueue.main.async {
-                    deferredIntentContext.completion(.failed(error: error))
-                }
-            }
-        }
-    }
-
-    private static func fetchIntentClientSecretFromMerchant(intentConfig: IntentConfiguration,
-                                                            paymentMethodID: String,
-                                                            shouldSavePaymentMethod: Bool) async throws -> String {
-      try await withCheckedThrowingContinuation { continuation in
-
-          if let confirmHandlerForServerSideConfirmation = intentConfig.confirmHandlerForServerSideConfirmation {
-              confirmHandlerForServerSideConfirmation(paymentMethodID, shouldSavePaymentMethod, { result in
-                  continuation.resume(with: result)
-              })
-          } else if let confirmHandler = intentConfig.confirmHandler {
-              confirmHandler(paymentMethodID, { result in
-                  continuation.resume(with: result)
-              })
-          }
-      }
-    }
 }
 
 /// A helper method to compare shipping details
